requirements.txt
apturl==0.5.2
asgiref==3.2.10
asn1crypto==0.24.0
command-not-found==0.3
config==0.5.0.post0
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap3==12.1.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.0
httplib2==0.14.0
MarkupSafe==1.1.0
Pillow==7.0.0
psycopg2==2.8.5
python-dateutil==2.7.3
python-decouple==3.3
pytz==2019.3
pyuploadcare==2.7.0
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
whitenoise==5.1.0
xkit==0.0.0
